/sbin/pup_event_frontend_d: Removed unused code.
Check for partitions 1-4 on some occasions,
were it was just checking for partition 1; since a disk
may have up to 4 primary partitions, but partition 1 may be deleted
and some of the other partition numbers may be still available.

Add some more vebosity messages.

Do not offer fsck option to fat and ntfs filesystems in MOREITEMS.