/sbin/pup_event_frontend_d: Huge code cleanup. Remove unused code.
Omitt check for icon already on that position, since the old code only checked
in the 10th range , which would only see nearby icons.
Icons are 48 pixels, and max. PIN_GRID_STEP is 32 .
The code would need to check up to 4 10th for nearby icons.
The code does not work for STEPS 16 and 32 because they are more than 10 ,
it works for nearby icons with STEP 2 .

Some improvements for the JWM Tray hight .