/etc/rc.d/rc.shutdown: Sorry, quiet some BUGFIXES related to creating a savefile at shutdown.
1) Somehow initrd failed to write DEV1FS to PUPSTATE
   (which I need to look further into)
   and xDEVFS was still unset after DEV1FS got set 'xxx' .
   This resulted in no savefile dialogs at all and shutdown
   without saving.

2) Busybox du does not recognize the -B blocksize option.
   Therefore the test for savefile sizes -lt always failed.

3) While renaming the ACHOICE variable name to aCHOICE
   I forgot or overlooked to rename almost half dozen occurences
   of it, probably to make the diff look nicer and committing later.
   This either resulted in SIZEPFILE being unset and defaulting to 1 MB later,
   or greping missing fields greped wrong sizes.
   The savefile was either 1024 or 1280 KB always.
   And that is far too little.

Other bugfixes:
1)
get_lineno() got not passed the LINEP , but echo it seems.
Now using semicolon, to make LINEP global.

2)
Some test -ABC on VARIABLE only was missing double quotes.
This may result on wrong return values if VARIABLE failed being set.

Add some $Q $VERB to modprobe.

TODOS :
1)
_continue() function does not work somehow to write and read to from /dev/console .
It is called in case of verbose parameter passed to rc.shutdown .

2)
Somehow it failed to read the password a second time from console in case of light encryption.
But the savefile got created anyway, and needs password to be mounted.
So am not sure if it really needs a second time.