/etc/rc.d/rc.launchxwin2: Added modesetting code from Fox3-Dell755 branch. Added one function _load_driver() to it to load the driver, if not already loaded. Currently only for i915 intel driver. Will not have the HW to test for Nvidia and ATI graphics.
Added a sleep 1 to let launchxwin chvt 1
and a chvt 2 before the modesetting code,
because modesetting does only work if it is done in the
current virtual terminal apparently.

Added stub to not modeset if kernelparameters
"nomodeset" or "vesa" are found in /proc/cmdline.
Will need to check the kernelparameters.txt if these
are already used by the kernel.