/usr/sbin/xorgwizard: Tidy up the dialog texts. Try to remember not only older card0driver, but also   VertRefresh and    HORIZSYNC lines in xorg.conf after editing it in mp editor. What to do with tis, I am not sure for now, since i have to figure out, how the VertRefresh and HORIZSYNC values are created by Xorg -configure or _do_ddcprobe() .
Now at a practically unbound framebuffer vga-console
:
bash-3.00# lsmod | grep fb
nvidiafb               30076  0
fb_ddc                   904  1 nvidiafb

the Xorg driver 'vesa' refuses to load claiming,
the console would be bound.

"modesetting" driver shows up as first Card0 ( as usual ), but does not load
at the test since no /dev/dri/card0 node exists at kernel 2.6.32.y .
"nv" nv_drv.so does not get accepted by Xorg -configure if it is a symlink to
nv-2.1.15_drv.so, and even Xorg -showopts refuses the link and the
renamed link-target.
"nouveau" is accepted as both Card1 and Card2 after Card0
on this real-4-Core-AMD-AthlonII-X4, but wants the nouveau.ko kernel driver,
which was first introduced with kernel-2.6.34 as a staging driver.
When "nv" is renamed as a regular file it behaves like the i810OLD_drv.so:
It gets recognized, but the later newer driver completely ignored
( in the i810 case the intel and in the nv case the nouveau _drv.so Xorg driver) .