/sbin/init: Split the _get_errflg() function into two more functions: _get_dump_ext234() to get the dumpe2fs variables and process the year/month/day numbers _eval_dump() to compare the above processed numbers
While at it, detected some logical flaws in regards to future/past comparision.
My guess is, that the New Year switch confused me that time.
This New Year switch is absolutely no special problem.
If the scheduled check would have been in the past year(s): do fsck
If the Month Number is lesser than the Check Month Number, then if
the next check year is larger than the current year, the day number
is of no interest.